Put SAC-related tests in their own classes
References #46, rabbitmq/rabbitmq-server#3753
